[FE] refactor : Types 파일 통합 처리
