+++ frontend/src/service/review/pages/ReviewPage/index.tsx
@@ -3,8 +3,7 @@ import { Navigate, useParams, useNavigate, Link, useLocation } from 'react-route
 
 import cn from 'classnames';
 
-import { Question } from 'service/@shared/types/review';
-import { RedirectState } from 'service/@shared/types/review';
+import { Question, RedirectState } from 'service/@shared/types';
 
 import useSnackbar from 'common/hooks/useSnackbar';
 import useQuestions from 'service/review/hooks/useQuestions';
